% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual-cycle-metrics.R
\name{annual_cycle_metrics}
\alias{annual_cycle_metrics}
\title{Calculate annual cycle metrics for a raster time series}
\usage{
annual_cycle_metrics(x)
}
\arguments{
\item{x}{RasterStack with each layer corresponding to a time interval}
}
\value{
A RasterStack with four layers:
  - `max`: the maximum pixel value across all time intervals
  - `peak_interval`: the interval number associated with the maximum
  - `seasonality`: the seasonality or unevenness
  - `n_intervals`: the number of intervals with a non-NA pixel value
}
\description{
Calculate annual cycle metrics for a raster time series
}
\examples{
data(elephant_ud)
r <- annual_cycle_metrics(elephant_ud)
}
