% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorist-package.R
\docType{package}
\name{colorist-package}
\alias{colorist}
\alias{colorist-package}
\title{colorist: Coloring Wildlife Distributions in Space-Time}
\description{
Color and visualize wildlife distributions in
    space-time using raster data. In addition to enabling display of
    sequential change in distributions through the use of small multiples,
    'colorist' provides functions for extracting several features of
    interest from a sequence of distributions and for visualizing those
    features using HCL (hue-chroma-luminance) color palettes. Resulting
    maps allow for "fair" visual comparison of intensity values (e.g.,
    occurrence, abundance, or density) across space and time and can be
    used to address questions about where, when, and how consistently a
    species, group, or individual is likely to be found.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mstrimas/colorist}
  \item Report bugs at \url{https://github.com/mstrimas/colorist/issues}
}

}
\author{
\strong{Maintainer}: Matthew Strimas-Mackey \email{mes335@cornell.edu} (\href{https://orcid.org/0000-0001-8929-7776}{ORCID})

Authors:
\itemize{
  \item Justin Schuetz \email{justinschuetz@gmail.com} (\href{https://orcid.org/0000-0002-6163-538X}{ORCID})
  \item Tom Auer \email{mta45@cornell.edu} (\href{https://orcid.org/0000-0001-8619-7147}{ORCID})
}

}
\keyword{internal}
