% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elephant_ud}
\alias{elephant_ud}
\title{African Elephant utilization distribution}
\format{
An object of class \code{RasterStack} of dimension 218 x 220 x 4.
}
\source{
\url{https://www.datarepository.movebank.org/handle/10255/move.812}
}
\usage{
elephant_ud
}
\description{
A \link[raster:RasterStack-class]{RasterStack} of the \href{https://en.wikipedia.org/wiki/Utilization_distribution}{utilization distribution} for a
single African Elephant in Etosha National Park. Cell values represent the
probability density that the individual elephant was found at a given
location and the four layers represent four successive years of data. This
distribution was derived from tracking data using the \code{adehabitatHR} package.
derived from tracking data.
}
\references{
Tsalyuk M., W. Kilian, B. Reineking, W. Marcus. 2018. Temporal variation in
resource selection of African elephants follows long term variability in
resource availability. Ecological Monographs.
\url{https://doi.org/10.1002/ecm.1348}

W. Kilian, W.M. Getz, R. Zidon, M. Tsalyuk. 2018. Data from: Temporal
variation in resource selection of African elephants follows long term
variability in resource availability. Movebank Data Repository.
\url{https://doi.org/10.5441/001/1.3nj3qj45}
}
\keyword{datasets}
