% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fiespa_occ}
\alias{fiespa_occ}
\title{eBird Status & Trends Field Sparrow occurrence probability}
\format{
An object of class \code{RasterStack} of dimension 100 x 121 x 12.
}
\source{
\url{https://ebird.org/science/status-and-trends}
}
\usage{
fiespa_occ
}
\description{
A \link[raster:Raster-classes]{RasterStack} of the expected probability of
occurrence of Field Sparrow from the \href{https://ebird.org/science/status-and-trends}{eBird Status & Trends} project. Each of the 12
layers in the stack represent the estimated occurrence for a given month of
the year over a regular grid of points covering the full range of the
species. To reduce file size, these data have been aggregated from the native
3 km spatial resolution and weekly temporal resolution to monthly, 27 km
resolution.
}
\details{
For further details on these data, and to access the data for more species,
consult the \href{https://ebird.github.io/ebirdst/}{documentation for the \code{ebirdst} package}.
}
\references{
Fink, D., T. Auer, A. Johnston, M. Strimas-Mackey, S. Ligocki,
O. Robinson, W. Hochachka, L. Jaromczyk, C. Crowley, K. Dunham,
A. Stillman, I. Davies, A. Rodewald, V. Ruiz-Gutierrez, C. Wood. 2023.
eBird Status and Trends, Data Version: 2022; Released: 2023.
Cornell Lab of Ornithology, Ithaca, New York.
\url{doi:10.2173/ebirdst.2022}
}
\keyword{datasets}
