% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend_groups}
\alias{legend_groups}
\title{Make HCL legend for the distributions of distinct groups}
\usage{
legend_groups(x, specificity = TRUE, group_labels = NULL, return_df = FALSE)
}
\arguments{
\item{x}{RasterStack or integer giving the number of layers to build the
HCL legend for.}

\item{specificity}{logical indicating whether to visualize the variation in
intensity in the legend or to fix intensity at 100.}

\item{group_labels}{character vector with two elements to be used as labels
for the start and end points of the time axis (i.e. x-axis) in the legend.}

\item{return_df}{logical indicating whether to return the legend as a
\code{ggplot2} object or return a data frame containing the necessary data to
build the legend.}
}
\value{
A ggplot2 plot object of the legend. Alternatively,
\code{return_df = TRUE} will return a data frame containing a data frame
containing the data needed to build the legend. The data frame columns are:
\itemize{
\item \code{specificity}: amount of variation between layers; mapped to chroma.
\item \code{layer}: layer with the maximum cell value; mapped to hue.
\item \code{color}: color associated with the given specificity and peak layer.
\item \code{intensity}: maximum cell value across layers divided by the maximum
value across all layers and cells; mapped to alpha level.
}
}
\description{
Make HCL legend for the distributions of distinct groups
}
\examples{
# load elephant data
data(elephant_ud)
# treat layers as groups of elephants, not years for a single elephant
names(elephant_ud) <- paste0("group", 1:4)

# generate hcl legend
legend_groups(elephant_ud)
}
\seealso{
\link{legend_timecycle} for cyclical sequences of distributions and
\link{legend_timeline} for linear sequences of distributions.

Other legend: 
\code{\link{legend_timecycle}()},
\code{\link{legend_timeline}()}
}
\concept{legend}
