% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend_timecycle}
\alias{legend_timecycle}
\title{Make HCL legend for a cyclical sequence of distributions}
\usage{
legend_timecycle(x, specificity = TRUE, return_df = FALSE)
}
\arguments{
\item{x}{RasterStack or integer giving the number of layers to build the
HCL legend for.}

\item{specificity}{logical indicating whether to show a single color wheel or
three wheels demonstrating the palette for intensity values of 0, 50, and
100.}

\item{return_df}{logical indicating whether to return the legend as a
\code{ggplot2} object or return a data frame containing the necessary data to
build the legend.}
}
\value{
A ggplot2 plot object of the legend. Alternatively,
\code{return_df = TRUE} will return a data frame containing a data frame
containing the data needed to build the legend. The data frame columns are:
\itemize{
\item \code{specificity}: amount of variation between layers; mapped to chroma.
\item \code{layer}: layer with the maximum cell value; mapped to hue.
\item \code{color}: color associated with the given specificity and peak layer.
\item \code{intensity}: maximum cell value across layers divided by the maximum
value across all layers and cells; mapped to alpha level.
}
}
\description{
Make HCL legend for a cyclical sequence of distributions
}
\examples{
# load field sparrow data
data(fiespa_occ)

# generate hcl legend
legend_timecycle(fiespa_occ)
}
\seealso{
\link{legend_timeline} for linear sequences of distributions.

Other palette: 
\code{\link{legend_timeline}()},
\code{\link{palette_groups}()},
\code{\link{palette_timecycle}()},
\code{\link{palette_timeline}()}
}
\concept{palette}
