% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend_timeline}
\alias{legend_timeline}
\title{Make HCL legend for a linear sequence of distributions}
\usage{
legend_timeline(
  x,
  start_hue = -130,
  specificity = TRUE,
  time_labels = NULL,
  return_df = FALSE
)
}
\arguments{
\item{x}{RasterStack or integer giving the number of layers to build the
HCL legend for.}

\item{start_hue}{integer between -360 and 360 representing the starting hue
in the color wheel. The ending hue will be \code{start_hue} + 180. For
further details, consult the documentation for \link[colorspace:rainbow_hcl]{colorspace::rainbow_hcl}.
Recommended values are -130, giving a blue-pink-yellow palette, and 50,
giving a yellow-green-blue palette.}

\item{specificity}{logical indicating whether to show a single color wheel or
three wheels demonstrating the palette for intensity values of 0, 50, and
100.}

\item{time_labels}{character vector with two elements to be used as labels
for the start and end points of the time axis (i.e. x-axis) in the legend.}

\item{return_df}{logical indicating whether to return the legend as a
\code{ggplot2} object or return a data frame containing the necessary data to
build the legend.}
}
\value{
A ggplot2 plot object of the legend. Alternatively,
\code{return_df = TRUE} will return a data frame containing a data frame
containing the data needed to build the legend. The data frame columns are:
\itemize{
\item \code{specificity}: amount of variation between layers; mapped to chroma.
\item \code{layer}: layer with the maximum cell value; mapped to hue.
\item \code{color}: color associated with the given specificity and peak layer.
\item \code{intensity}: maximum cell value across layers divided by the maximum
value across all layers and cells; mapped to alpha level.
}
}
\description{
Make HCL legend for a linear sequence of distributions
}
\examples{
# load field sparrow data
data(fiespa_occ)

# generate hcl legend
legend_timeline(fiespa_occ)
}
\seealso{
\link{palette_timecycle} for cyclical sequences of distributions.

Other palette: 
\code{\link{legend_timecycle}()},
\code{\link{palette_groups}()},
\code{\link{palette_timecycle}()},
\code{\link{palette_timeline}()}
}
\concept{palette}
