% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-hcl-palette.R
\name{make_hcl_palette}
\alias{make_hcl_palette}
\title{Genreate a stack of HCL color wheels}
\usage{
make_hcl_palette(x)
}
\arguments{
\item{x}{RasterStack or integer giving the number of intervals.}
}
\value{
A data frame with the following columns:
  - `seasonality`: the seasonality, which is mapped to chroma
  - `interval`: the time interval, which is mapped to hue
  - `color`: the hex color associated with the given seasonality and interval
}
\description{
Genreate a stack of HCL color wheels
}
\examples{
data(elephant_ud)
pal <- make_hcl_palette(elephant_ud)
head(pal)
}
