% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_multiples}
\alias{map_multiples}
\title{Visualize distributions as a single map}
\usage{
map_multiples(
  x,
  palette,
  ncol,
  lambda = 0,
  labels = NULL,
  suppress_map = FALSE
)
}
\arguments{
\item{x}{RasterStack of distributions processed by \code{\link[=metrics_pull]{metrics_pull()}}.}

\item{palette}{data frame containing an HCL color palette.}

\item{ncol}{integer specifying the number of columns in the grid of plots.}

\item{lambda}{integer controlling the offset in the Box-Cox transformation of
intensities to alpha levels via the \code{\link[scales:modulus_trans]{scales::modulus_trans()}} function. Use
0 for Box-Cox type 1 or any non-negative number for Box-Cox type 2.}

\item{labels}{character vector of layer labels for each plot. The default is
to not show labels.}

\item{suppress_map}{logical controlling whether the map should be printed
only (TRUE) or printed and returned.}
}
\value{
A ggplot2 plot object.
}
\description{
Visualize distributions as a single map
}
\examples{
data("elephant_ud")

# prepare data
r <- metrics_pull(elephant_ud)

# generate palette
pal <- palette_timeline(elephant_ud)

# produce map
map_multiples(x, palette)
}
