% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-to-colors.R
\name{metrics_to_colors}
\alias{metrics_to_colors}
\title{Assign cell colors and alpha levels based on annual cycle metrics}
\usage{
metrics_to_colors(x, palette)
}
\arguments{
\item{x}{RasterStack of seasonility metrics from [annual_cycle_metrics()]}

\item{palette}{data frame containing an HCL color palette as output from
[make_hcl_palette()]}
}
\value{
A RasterStack with four layers, specifying the RGB color and alpha
  level, all on the 0-1 scale:
  - `r`: red channel
  - `g`: green channel
  - `b`: blue channel
  - `alpha`: alpha level
}
\description{
Assign cell colors and alpha levels based on annual cycle metrics
}
\examples{
data(elephant_ud)
# generate annual cyle metrics and a color palette based on sample data
r <- annual_cycle_metrics(elephant_ud)
pal <- make_hcl_palette(elephant_ud)
# assign colors
rgb <- metrics_to_colors(r, pal)
}
