% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-to-hex.R
\name{metrics_to_hex}
\alias{metrics_to_hex}
\title{Assign cell colors and alpha levels based on annual cycle metrics}
\usage{
metrics_to_hex(x, palette)
}
\arguments{
\item{x}{RasterStack of seasonility metrics from [annual_cycle_metrics()]}

\item{palette}{data frame containing an HCL color palette as output from
[make_hcl_palette()]}
}
\value{
A data frame with four columns, specifying cell center (`x` and `y`),
  the color (`color`), and the alpha level (`alpha`).
}
\description{
Assign cell colors and alpha levels based on annual cycle metrics
}
\examples{
data(elephant_ud)
# generate annual cyle metrics and a color palette based on sample data
r <- annual_cycle_metrics(elephant_ud)
pal <- make_hcl_palette(elephant_ud)
# assign colors
cell_colors <- metrics_to_hex(r, pal)
head(cell_colors)
}
