% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{legend_timeline}
\alias{legend_timeline}
\title{Make HCL legend for a linear sequence of distributions}
\usage{
legend_timeline(
  palette,
  specificity = TRUE,
  time_labels = NULL,
  return_df = FALSE
)
}
\arguments{
\item{palette}{data frame containing a color palette generated by
\link{palette_timeline}.}

\item{specificity}{logical indicating whether to show a single legend or
three legend demonstrating the palette for specificity values of 0, 50, and
100.}

\item{time_labels}{character vector with two elements to be used as labels
for the start and end points of the time axis (i.e. x-axis) in the legend.}

\item{return_df}{logical indicating whether to return the legend as a
\code{ggplot2} object or return a data frame containing the necessary data to
build the legend.}
}
\value{
A ggplot2 plot object of the legend. Alternatively,
\code{return_df = TRUE} will return a data frame containing a data frame
containing the data needed to build the legend. The data frame columns are:
\itemize{
\item \code{specificity}: amount of variation between layers; mapped to chroma.
\item \code{layer_id}: integer identifying the layer containing the maximum
intensity value; mapped to hue.
\item \code{color}: color associated with the given specificity and peak layer.
\item \code{intensity}: maximum cell value across layers divided by the maximum
value across all layers and cells; mapped to alpha level.
}
}
\description{
Make HCL legend for a linear sequence of distributions
}
\examples{
# load elephant data
data(elephant_ud)

# generate hcl palette
pal <- palette_timeline(elephant_ud)

# create legend for palette
legend_timeline(pal)
}
\seealso{
\link{legend_timecycle} for cyclical sequences of distributions and
\link{legend_groups} for distributions of distinct groups.

Other legend: 
\code{\link{legend_groups}()},
\code{\link{legend_timecycle}()}
}
\concept{legend}
