% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_multiples}
\alias{map_multiples}
\title{Visualize distributions as a single map}
\usage{
map_multiples(x, palette, ncol, lambda = 0, labels = NULL, return_df = FALSE)
}
\arguments{
\item{x}{RasterStack of distributions processed by \code{\link[=metrics_pull]{metrics_pull()}}.}

\item{palette}{data frame containing an HCL color palette.}

\item{ncol}{integer specifying the number of columns in the grid of plots.}

\item{lambda}{integer controlling the offset in the Box-Cox transformation of
intensities to alpha levels via the \code{\link[scales:modulus_trans]{scales::modulus_trans()}} function. Use
0 for Box-Cox type 1 or any non-negative number for Box-Cox type 2.}

\item{labels}{character vector of layer labels for each plot. The default is
to not show labels.}

\item{return_df}{logical specifying whether the function should return a
ggplot2 plot object (FALSE) or a data frame containing the raster data and
associated cell colors.}
}
\value{
A ggplot2 plot object of the map. Alternatively, \code{return_df = TRUE}
will return a data frame containing the raster data in data frame format
along with the associated cell colors. The data frame columns are:
\itemize{
\item \code{x},\code{y}: coordinates of raster cell centers.
\item \code{cell_number}: integer indicating the cell number.
\item \code{layer_cell}: a unique ID for the cell within the layer in the format:
\code{"layer-cell_number"}.
\item \code{intensity}: cell value divided by the maximum in the layer; mapped to
alpha level.
\item \code{specificity}: amount of variation between layers; mapped to chroma.
\item \code{layer}: raster layer that this value came from; mapped to hue.
\item \code{color}: color associated with the given specificity and peak layer.
}
}
\description{
Visualize distributions as a single map
}
\examples{
# load elephant data
data("elephant_ud")

# prepare data
r <- metrics_pull(elephant_ud)

# generate palette
pal <- palette_timeline(elephant_ud)

# produce map
map_multiples(r, pal)
}
\seealso{
Other map: 
\code{\link{map_single}()}
}
\concept{map}
