% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_single}
\alias{map_single}
\title{Visualize distributions as a single map}
\usage{
map_single(x, palette, layer, lambda = 0, suppress_map = FALSE)
}
\arguments{
\item{x}{RasterStack of distributions processed by \code{\link[=metrics_pull]{metrics_pull()}} or
\code{\link[=metrics_distill]{metrics_distill()}}.}

\item{palette}{data frame containing an HCL color palette.}

\item{layer}{integer or character corresponding to the index or name of the
single layer to map. This argument is ignored if \code{\link[=metrics_distill]{metrics_distill()}} was
used.}

\item{lambda}{integer controlling the offset in the Box-Cox transformation of
intensities to alpha levels via the \code{\link[scales:modulus_trans]{scales::modulus_trans()}} function. Use
0 for Box-Cox type 1 or any non-negative number for Box-Cox type 2.}

\item{suppress_map}{logical controlling whether the map should be printed
only (TRUE) or printed and returned.}
}
\value{
A ggplot2 plot object.
}
\description{
Visualize distributions as a single map
}
\examples{
data("elephant_ud")

# prepare metrics
r <- metrics_distill(elephant_ud)

# generate palette
pal <- palette_timeline(elephant_ud)

# produce map
map_single(x, palette)
}
