% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_distill}
\alias{metrics_distill}
\title{Distill a raster stack into a set of metrics for each cells}
\usage{
metrics_distill(x)
}
\arguments{
\item{x}{RasterStack with each layer typically corresponding to a time
interval.}
}
\value{
A RasterStack with four layers:
\itemize{
\item \code{intensity_max}: the maximum pixel intensity across all layers
\item \code{intensity_mean}: the mean pixel intensity across all layers
\item \code{peak_layer}: the layer number associated with the maximum intensity.
\item \code{specificity}: the amount of variation between layers for each cell, this
is a measure of seasonality if the layers represents different times of
year.
\item \code{n_layers}: the number of intervals with a non-NA pixel value
}

Here intensity is the result of dividing all cell values by the maximum
cell value across all layers. This maximum value is stored as the
\code{"maximum"} attribute.
}
\description{
Distill a raster stack into a set of metrics for each cells
}
\examples{
library(raster)
data("elephant_ud")

r <- metrics_distill(elephant_ud)
print(r)
attr(r, "maximum")
}
\seealso{
Other metrics: 
\code{\link{metrics_pull}()}
}
\concept{metrics}
