% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_pull}
\alias{metrics_pull}
\title{Transform a raster stack to intensity values}
\usage{
metrics_pull(x)
}
\arguments{
\item{x}{RasterStack with each layer typically corresponding to the
distribution of an individual or species at a given time interval.}
}
\value{
A RasterStack resulting from converting the input stack to intensity
values by dividing each layer by the maximum pixel value in that layer.
These layer maxima are stored as the \code{"maximum"} attribute.
}
\description{
Transform a raster stack to intensity values
}
\examples{
# load elephant data
data("elephant_ud")
r <- metrics_pull(elephant_ud)
print(r)
# maximum values for each layer stored as an attribute
attr(r, "maximum")
}
\seealso{
Other metrics: 
\code{\link{metrics_distill}()}
}
\concept{metrics}
