% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_groups}
\alias{palette_groups}
\title{Make an HCL palette for visualizing the distributions of distinct groups}
\usage{
palette_groups(x)
}
\arguments{
\item{x}{RasterStack or integer describing the number of layers for which
colors need to be generated.}
}
\value{
A data frame with three columns:
\itemize{
\item \code{layer_id}: integer identifying the layer containing the maximum
intensity value; mapped to hue.
\item \code{specificity}: the degree to which intensity values are unevenly
distributed across layers; mapped to chroma.
\item \code{color}: the hexadecimal color associated with the given layer and
specificity values.
}
}
\description{
This function generates an HCL palette for visualizing
a small set of distributions (e.g., utilization
distributions describing space use by five individuals or
species distributions for five separate species).
}
\examples{
# load elephant data
# treat layers as groups of elephants, not years for a single elephant
data(elephant_ud)

# generate hcl color palette
pal <- palette_groups(elephant_ud)
head(pal)

# visualize the palette in HCL space  with colorspace::hclplot
library(colorspace)
hclplot(pal[pal$specificity == 100, ]$color)
}
\seealso{
\link{palette_timecycle} for cyclical sequences of distributions and
\link{palette_timeline} for linear sequences of distributions.

Other palette: 
\code{\link{palette_timecycle}()},
\code{\link{palette_timeline}()}
}
\concept{palette}
