% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_timecycle}
\alias{palette_timecycle}
\title{Make HCL palette for a cyclical sequence of distributions}
\usage{
palette_timecycle(x)
}
\arguments{
\item{x}{RasterStack or integer giving the number of layers to build the
color palette for.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{specificity}: seasonality; mapped to chroma.
\item \code{peak_layer}: layer with the maximum cell value; mapped to hue.
\item \code{color}: ex color associated with the given specificity and peak layer.
}
}
\description{
Make HCL palette for a cyclical sequence of distributions
}
\examples{
data(fiespa_occ)
pal <- palette_timecycle(fiespa_occ)
head(pal)

# visualize the palette in HCL space with colorspace::hclplot
library(colorspace)
hclplot(pal[pal$specificity == 50, ]$color)
}
\seealso{
\link{palette_timeline} for linear sequences of distributions and
\link{palette_groups} for distributions of distinct groups.

Other palette: 
\code{\link{palette_groups}()},
\code{\link{palette_timeline}()}
}
\concept{palette}
