% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_timeline}
\alias{palette_timeline}
\title{Make HCL palette for a linear sequence of distributions}
\usage{
palette_timeline(x, start_hue = -130)
}
\arguments{
\item{x}{RasterStack or integer giving the number of layers to build the
color palette for.}

\item{start_hue}{integer between -360 and 360 representing the starting hue
in the color wheel. The ending hue will be \code{start_hue} + 180. For
further details, consult the documentation for \link[colorspace:rainbow_hcl]{colorspace::rainbow_hcl}.
Recommended values are -130, giving a blue-pink-yellow palette, and 50,
giving a yellow-green-blue palette.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{specificity}: seasonality; mapped to chroma.
\item \code{peak_layer}: layer with the maximum cell value; mapped to hue.
\item \code{color}: ex color associated with the given specificity and peak layer.
}
}
\description{
Make HCL palette for a linear sequence of distributions
}
\examples{
data(fiespa_occ)
pal_a <- palette_timeline(fiespa_occ)
head(pal_a)

# try a different starting hue
pal_b <- palette_timeline(fiespa_occ, start = 50)

# visualize the palette in HCL space  with colorspace::hclplot
library(colorspace)
hclplot(pal_a[pal_a$specificity == 50, ]$color)
hclplot(pal_a[pal_b$specificity == 50, ]$color)
}
\seealso{
\link{palette_timecycle} for cyclical sequences of distributions and
\link{palette_groups} for distributions of distinct groups.

Other palette: 
\code{\link{palette_groups}()},
\code{\link{palette_timecycle}()}
}
\concept{palette}
